+++ test/src/analyzers/lint_analyzer/rules/rules_list/avoid_unused_parameters/examples/tear_off_example.dart
@@ -1,0 +1,20 @@
+class TestClass {
+  String value;
+
+  void _someMethod(Function(String, int) callback) {
+    callback();
+  }
+
+  void _otherMethod(String string, int value) {
+    print(string);
+  }
+
+  // LINT
+  void _anotherMethod(String firstString, String secondString) {
+    someMethod(_otherMethod);
+    _someOtherMethod(value);
+  }
+
+  // LINT
+  _someOtherMethod(String value) {}
+}